+++ src/components/IIngredientDetails/IngredientDetails.tsx
