--- src/types.ts
@@ -1,19 +1,0 @@
-export interface IData {
-  _id?: string,
-  name: string,
-  type: string,
-  proteins: number,
-  fat: number,
-  carbohydrates: number,
-  calories: number,
-  price: number,
-  image: string
-  image_mobile: string,
-  image_large: string,
-  v?: number
-}
-
-
-export type TCardElement = {
-  data: IData[];
-}
